rename modules, mark as private
